Changes:- 1) Moved Gateway from entity management to iot-manager 2) Moved Gateway Schema and Model from entity management to iot-manager 3) Moved gateways heirarchy fetch from project to gateway model.
