+++ apps/acqdat_api/lib/acqdat_api_web/plugs/load_gateway.ex
@@ -1,6 +1,6 @@
 defmodule AcqdatApiWeb.Plug.LoadGateway do
   import Plug.Conn
-  alias AcqdatCore.Model.EntityManagement.Gateway, as: GModel
+  alias AcqdatCore.Model.IotManager.Gateway, as: GModel
 
   @spec init(any) :: any
   def init(default), do: default
