+++ apps/acqdat_api/lib/acqdat_api_web/validators/iot-manager/gateway.ex
@@ -1,4 +1,4 @@
-defmodule AcqdatApiWeb.Validators.EntityManagement.Gateway do
+defmodule AcqdatApiWeb.Validators.IotManager.Gateway do
   use Params
 
   defparams(
