+++ apps/acqdat_api/lib/acqdat_api_web/views/entity-management/entity_view.ex
@@ -1,7 +1,7 @@
 defmodule AcqdatApiWeb.EntityManagement.EntityView do
   use AcqdatApiWeb, :view
   alias AcqdatApiWeb.EntityManagement.ProjectView
-  alias AcqdatApiWeb.EntityManagement.GatewayView
+  alias AcqdatApiWeb.IotManager.GatewayView
 
   def render("organisation_tree.json", %{org: org}) do
     %{
