+++ apps/acqdat_api/test/acqdat_api_web/controllers/iot-manager/gateway_controller_test.exs
@@ -1,4 +1,4 @@
-defmodule AcqdatApiWeb.EntityManagement.GatewayControllerTest do
+defmodule AcqdatApiWeb.IotManager.GatewayControllerTest do
   use ExUnit.Case, async: true
   use AcqdatApiWeb.ConnCase
   use AcqdatCore.DataCase
